==> notebooks/01_cleaning_500_albums.ipynb <==
## Import libraries

import pandas as pd
import numpy as np

import warnings
warnings.filterwarnings('ignore')

pd.set_option('display.max_columns', None)  
#pd.set_option('display.max_rows', None)  

## Import dataset

data = pd.read_csv('../data/raw/DB_Kaggle_top500albums.csv', encoding='latin1')

data.head(3)

## Clean data

#ensure there are not spaces 

data['Album'] = data['Album'].apply(lambda x: x.strip() if isinstance(x, str) else x)

data['Artist'] = data['Artist'].apply(lambda x: x.strip() if isinstance(x, str) else x)

#rename the columns 

data.columns

columns = ['toprankingorder', 'yearrs', 'albumnamers', 'artistnamers', 'genrers', 'subgenrers']

data.columns = columns

data.head(3)

#check datatype

data.info()

data = data.iloc[:, [0, 2, 3,1]]

data.head()

data.to_csv('../data/clean/DB_Kaggle_top500albums_clean.csv', index=False)
