==> notebooks/02_selenium_50horriblealbums.ipynb <==
## Import libraries

from selenium import webdriver
import pandas as pd
from selenium.webdriver.common.by import By
import time 
from selenium.webdriver.chrome.options import Options
from webdriver_manager.chrome import ChromeDriverManager

#driver configuration
opciones=Options()

opciones.add_experimental_option('excludeSwitches', ['enable-automation'])
opciones.add_experimental_option('useAutomationExtension', False)
opciones.headless=False    # si True, no aperece la ventana (headless=no visible)
opciones.add_argument('--start-maximized')         # comienza maximizado
#opciones.add_argument('user-data-dir=selenium')    # mantiene las cookies
#opciones.add_extension('driver_folder/adblock.crx')       # adblocker
opciones.add_argument('--incognito')

PATH = 'driver/chromedriver'

driver = webdriver.Chrome()

driver.get('https://www.google.es')

## Extract and transform data

url = 'https://www.rollingstone.com/music/music-lists/horrible-albums-by-brilliant-artists-1234672895/'

driver.get(url)

#cookies 

rejectcookies = driver.find_element(By.XPATH,'//*[@id="onetrust-reject-all-handler"]')
rejectcookies.click()

#ranking number 

rankingnumber = driver.find_elements(By.CLASS_NAME,'c-gallery-vertical-album__number')
len(rankingnumber)

rankingnumber = [item.text for item in driver.find_elements(By.CLASS_NAME,'c-gallery-vertical-album__number')]
rankingnumber[0:5]

#band name  

bandname = driver.find_elements(By.CLASS_NAME,'c-gallery-vertical-album__title')
len(bandname)

bandname = [item.text for item in driver.find_elements(By.CLASS_NAME,'c-gallery-vertical-album__title')]
bandname[0:5]

#albumname  

albumname = driver.find_elements(By.CLASS_NAME,'c-gallery-vertical-album__additional--description')
len(albumname)

albumname = [item.text.replace("'","") for item in driver.find_elements(By.CLASS_NAME,'c-gallery-vertical-album__additional--description')]
albumname[0:5]

#releaseyear 

releaseyear = driver.find_elements(By.CLASS_NAME,'c-gallery-vertical-album__subtitle-1')
len(releaseyear)

releaseyear = [item.text for item in driver.find_elements(By.CLASS_NAME,'c-gallery-vertical-album__subtitle-1')]
releaseyear[0:5]

#rationale

rationale = driver.find_elements(By.CLASS_NAME,'c-gallery-vertical-album__description')
len(rationale)

rationale = [item.text for item in driver.find_elements(By.CLASS_NAME,'c-gallery-vertical-album__description')]
rationale[0]

data = {'rshorriblerankingorder': rankingnumber, 'rsartist': bandname, 'rsalbum': albumname, 'rsyear': releaseyear, 'rsrationale': rationale}

finaltable = pd.DataFrame(data)
finaltable.head()

#review datatype 

finaltable.info()

finaltable['rshorriblerankingorder'] = finaltable['rshorriblerankingorder'].astype('int64')

finaltable['rshorriblerankingorder'].dtype

finaltable.shape

finaltable = finaltable.iloc[:, [0, 2, 1, 4, 3]] 

finaltable.head()

finaltable.info()

finaltable.to_csv('../data/clean/WS_50horriblealbums_clean.csv', index=False)
