==> notebooks/03_albumstargetlist.ipynb <==
### Importing libraries

import pandas as pd
import numpy as np
import time 

import warnings
warnings.filterwarnings('ignore')

pd.set_option('display.max_columns', None)  
#pd.set_option('display.max_rows', None)  

### What albums am I interested on?

#### Rolling Stone magazine Top 500

topalbums = pd.read_csv('../data/clean/DB_Kaggle_top500albums_clean.csv', encoding='latin1')

topalbums.head(3)

#### Rolling Stone article Horrible 50

horriblealbums = pd.read_csv('../data/clean/WS_50horriblealbums_clean.csv', encoding='latin1')

horriblealbums.head(3)

#### Full album list 

#need to be conscious of the fact that some albums have the same name 

topalbums['rsalbum'].value_counts().head(3)

#let us get unique values for album_artist 

topalbums['album_artist'] = topalbums['rsalbum'] + ' - ' + topalbums['rsartist'] 

topalbums.head(3)

horriblealbums['rsalbum'].value_counts().head(2)

horriblealbums['album_artist'] = horriblealbums['rsalbum'] + ' - ' + horriblealbums['rsartist']

horriblealbums.head(2)

all_albums = pd.concat([topalbums, horriblealbums], ignore_index=True)
all_albums.shape

all_albums.head(3)

all_albums.to_csv('../data/clean/all_550_albums_clean.csv', index=False)
