==> notebooks/04_api_albums.ipynb <==
## Import libraries

import pandas as pd
import numpy as np
import time 

import warnings
warnings.filterwarnings('ignore')

pd.set_option('display.max_columns', None)  
#pd.set_option('display.max_rows', None)  

## Get artist target list 

all_albums = pd.read_csv('../data/clean/all_550_albums_clean.csv', encoding='latin1')

all_albums.head(2)

len(all_albums['rsartist'].unique())

## Spotify

#%pip install pyarrow
#%pip install joblib
#%pip install tqdm
#%pip install spotipy
#%pip install python-dotenv
#%pip install requests 

import requests as rq
import pyarrow
from joblib import Parallel, delayed
from tqdm import tqdm
import pymongo
from passwords import *

### Generate token

# Obtener el TOKEN

AUTH_URL = "https://accounts.spotify.com/api/token"

creds = {
    "grant_type": "client_credentials",
    "client_id": Client_ID,
    "client_secret": Client_secret
}

response = rq.post(AUTH_URL, 
                         data=creds, 
                         headers={"Content-Type": "application/x-www-form-urlencoded"})


TOKEN = response.json()['access_token']

# Conexión con la API

url = "https://api.spotify.com/"

headers = {"Authorization": f'Bearer {TOKEN}'}

response = rq.get(url,headers=headers)
response

### Get artist information 

#we will do a keyword search 

all_albums['rsartist'].head(3)

#### Test to understand the API for the info I am interested on 

endpoint = 'https://api.spotify.com/v1/search'

artistname = 'The Beatles'
albumname = "Sgt. Pepper's Lonely Hearts Club Band"

params = {'q': f'album:{albumname} artist:{artistname}','type': 'album','market':'US', 'limit':'1','offset':'0'}

response = rq.get(endpoint, headers=headers, params=params)
response

data = response.json()
data.keys()

#getalbumid
data['albums']['items'][0]['id']

#getalbumname
data['albums']['items'][0]['name']

#artist name 
data['albums']['items'][0]['artists'][0]['name']

#artist id
data['albums']['items'][0]['artists'][0]['id']

#album release date
data['albums']['items'][0]['release_date']

#total tracks
data['albums']['items'][0]['total_tracks']

#### target lists 

#replacing spaces for '+' for the search query 

albumlist = [x.replace(' ', '+') for x in all_albums['rsalbum']]
albumlist[0:3]

artistlist = [x.replace(' ', '+') for x in all_albums['rsartist']]
artistlist[0:3]

#function to get info

data_list = []

def get_album_info(artist_name, album_name):
    try:
        endpoint = 'https://api.spotify.com/v1/search'
        params = {
            'q': f'album:{album_name} artist:{artist_name}',
            'type': 'album',
            'market': 'US',
            'limit': '1',
            'offset': '0',
        }
        response = rq.get(endpoint, headers=headers, params=params)

        if response.status_code == 200:
            data = response.json()
            if data['albums']['items']:
                album_info = data['albums']['items'][0]
                data_list.append({
                    'albumnamers': album_name,
                    'albumidspotify': album_info['id'],
                    'albumnamespotify': album_info['name'],
                    'artistnamespotify': album_info['artists'][0]['name'],
                    'artistidspotify': album_info['artists'][0]['id'],
                    'releasedate': album_info['release_date'],
                    'totaltracks': album_info['total_tracks']
                })
            else:
                data_list.append({
                    'albumnamers': album_name,
                    'albumidspotify': None,
                    'albumnamespotify': None,
                    'artistnamespotify': None,
                    'artistidspotify': None,
                    'releasedate': None,
                    'totaltracks': None
                })
                print(f"No album found for {album_name} by {artist_name}")
        else:
            print(f"Failed to retrieve album information for {album_name} by {artist_name}. Status code: {response.status_code}")
    except Exception as e:
        print(f"An error occurred: {e}")

#execute for the entire list 
for artist, album in zip(artistlist, albumlist):
    get_album_info(artist, album)

# Create a DataFrame
df = pd.DataFrame(data_list)
df.head(3)

df.shape

df['albumnamers'] = df['albumnamers'].apply(lambda x: x.replace('+', ' ') if isinstance(x, str) else x)

df.head(2)

#we were unable to find information for 87 albums

na_col = df.isnull().sum()
na_col

df.to_csv('../data/clean/api_albums_clean.csv', index=False)
