==> notebooks/08_tableau_table_500_best.ipynb <==
## Import libraries

import pandas as pd
import numpy as np

import warnings
warnings.filterwarnings('ignore')

pd.set_option('display.max_columns', None)  # me muestre todas las columnas
#pd.set_option('display.max_rows', None)  # me muestre todas las filas

## Import files

top500 = pd.read_csv('../data/clean/DB_Kaggle_top500albums_clean.csv', encoding='latin1')
top500.head(3)

top500.rename(columns={'album': 'albumnamers'}, inplace=True)
top500.columns

spotifyalbums = pd.read_csv('../data/clean/api_albums_clean.csv', encoding='latin1')
spotifyalbums.head(3)

## Merge tables

sqltable_top500 = pd.merge(top500, spotifyalbums[['albumnamers', 'albumidspotify']], how='left', on='albumnamers')

sqltable_top500.head(2)

sqltable_top500.shape

## Clean NAs and duplicates

#clean NAs (some albums available in the RS DB were not found in Spotify)
sqltable_top500.dropna(inplace=True)

sqltable_top500.shape

sqltable_top500.duplicated().any()

import re

# Iterate over each column in the DataFrame
for col in sqltable_top500.columns:
    sqltable_top500[col] = sqltable_top500[col].apply(lambda x: re.sub(r'[^\w\s]', '', x) if isinstance(x, str) else x)

# Display the updated DataFrame
sqltable_top500.head()

## Column names

sqltable_top500.columns

finalcolumns = ['rs_toprankingorder', 'rs_albumname', 'rs_artist', 'rs_year', 'sp_albumid']

sqltable_top500.columns = finalcolumns

sqltable_top500.head(1)

## Reoder columns

#sqltable_top500 = sqltable_top500.iloc[:, [0, 1, 2, 3, 4,5]]

sqltable_top500['albums_sp_albumid'] = sqltable_top500['sp_albumid']

sqltable_top500.head(1)

## Data type

sqltable_top500.info()

sqltable_top500['rs_year'] = pd.to_datetime(sqltable_top500['rs_year'], format='%Y')

sqltable_top500['rs_year'].dtype

## Export 

sqltable_top500.to_csv('../data/tableau/tableautable_top500.csv', index=True, encoding='latin1')
