==> notebooks/09_tableau_table_50_horrible.ipynb <==
## Import libraries

import pandas as pd
import numpy as np

import warnings
warnings.filterwarnings('ignore')

pd.set_option('display.max_columns', None)  # me muestre todas las columnas
#pd.set_option('display.max_rows', None)  # me muestre todas las filas

## Import files

horrible50 = pd.read_csv('../data/clean/WS_50horriblealbums_clean.csv', encoding='latin1')
horrible50.head(2)

horrible50.rename(columns={'rsalbum': 'albumnamers'}, inplace=True)
horrible50.columns

spotifyalbums = pd.read_csv('../data/clean/api_albums_clean.csv', encoding='latin1')
spotifyalbums.head(3)

## Merge tables

sqltable_horrible50 = pd.merge(horrible50, spotifyalbums[['albumnamers', 'albumidspotify']], how='left', on='albumnamers')

sqltable_horrible50.head(2)

sqltable_horrible50.shape

## Clean NAs and duplicates

#clean NAs (some albums available in the RS DB were not found in Spotify)
sqltable_horrible50.dropna(inplace=True)

sqltable_horrible50.shape

sqltable_horrible50.duplicated().any()

## Column names

sqltable_horrible50.columns

finalcolumns = ['rs_horriblerankingorder', 'rs_albumname', 'rs_artist', 'rationale', 'rs_year','sp_albumid']

sqltable_horrible50.columns = finalcolumns

sqltable_horrible50.head(1)

## Reoder columns

sqltable_horrible50 = sqltable_horrible50.iloc[:, [5, 0, 1, 2, 4]]
sqltable_horrible50.head(1)

sqltable_horrible50['album_sp_albumid'] = sqltable_horrible50['sp_albumid']

sqltable_horrible50.head(1)

## Data type

sqltable_horrible50.info()

sqltable_horrible50['rs_year'] = pd.to_datetime(sqltable_horrible50['rs_year'], format='%Y')

sqltable_horrible50['rs_year'].dtype

sqltable_horrible50.info()

## Export 

sqltable_horrible50.to_csv('../data/tableau/tableautable_horrible50.csv', index=True, encoding='latin1')
