==> notebooks/12_table_table_tracks.ipynb <==
## Import libraries

import pandas as pd
import numpy as np

import warnings
warnings.filterwarnings('ignore')

pd.set_option('display.max_columns', None)  # me muestre todas las columnas
#pd.set_option('display.max_rows', None)  # me muestre todas las filas

## Import files 

trackbasicinfo = pd.read_csv('../data/clean/api_tracks1_clean.csv', encoding='latin1')

trackbasicinfo.head(2)

trackmetrics =  pd.read_csv('../data/clean/api_alltracks_clean.csv', encoding='latin1')

trackmetrics.head(2)

## Merge tables

tracks = pd.merge(trackbasicinfo, trackmetrics[['track_id', 'danceability','energy','acousticness','valence','tempo','loudness']], how='left', on='track_id')

tracks.head(2)

tracks.drop(['artist_id'], axis=1, inplace=True)

tracks.head(1)

## Export file

tracks.to_csv('../data/tableau/tableautable_tracks.csv', index=False, encoding='latin1')
