==> notebooks/99_append_tracks_data.ipynb <==
## Import libraries

import pandas as pd
import numpy as np

import warnings
warnings.filterwarnings('ignore')

pd.set_option('display.max_columns', None)  # me muestre todas las columnas
#pd.set_option('display.max_rows', None)  # me muestre todas las filas

## Import files

tracks0_30 = pd.read_csv('../data/clean/api_tracks1_clean_0_30.csv', encoding='latin1')

tracks30_100 = pd.read_csv('../data/clean/api_tracks1_clean_30_100.csv', encoding='latin1')

tracks100_200 = pd.read_csv('../data/clean/api_tracks1_clean_100_200.csv', encoding='latin1')

tracks200_400 = pd.read_csv('../data/clean/api_tracks1_clean_200_400.csv', encoding='latin1')

tracks400_800 = pd.read_csv('../data/clean/api_tracks2_clean_400_800.csv', encoding='latin1')

tracks800_1300 = pd.read_csv('../data/clean/api_tracks2_clean_800_1300.csv', encoding='latin1')

## Append data

datasets = [tracks0_30, tracks30_100, tracks100_200, tracks200_400, tracks400_800, tracks800_1300]

alltracks = pd.concat(datasets, ignore_index=True)

alltracks.head()

alltracks = alltracks.drop(['popularity', 'instrumentalness'], axis=1)

alltracks.head()

alltracks = alltracks.reset_index(drop=True)

alltracks.shape

## Export file

alltracks.to_csv('../data/clean/api_alltracks_clean.csv', index=False)
